Modification and amelioration of some unit test for AOI classe
